* call things related_content instead of sidebars * build related_content from recipes and embed in element JSON
