+++ recipes/html-element.yaml
@@ -1,3 +1,4 @@
+related_content: /content/related_content/html.yaml
 body:
 - prose.short_description
 - meta.interactive_example?
